Improve question generation (#16)
* Start rework of question generation.

Keep list of available questions.
Select the random question from list.
Make sure that operands are not repeated with previous question.

* Improve control with previous question.

Since useState is asynchronously updated it's not easy to capture correct
value and pass it down. Therefore previous questions are stored in
questions generator so that they don't have to be passed.